+++ packages/@expo/cli/src/export/exportApp.ts
@@ -127,15 +127,22 @@ export async function exportAppAsync(
   const bundles: Partial<Record<Platform, BundleOutput>> = {};
   const domComponentAssetsMetadata: Partial<Record<Platform, PlatformMetadata['assets']>> = {};
 
-  const spaPlatforms = useServerRendering
-    ? platforms.filter((platform) => platform !== 'web')
-    : platforms;
+  const spaPlatforms =
+    // TODO: Support server and static rendering for server component exports.
+    useServerRendering && !devServer.isReactServerComponentsEnabled
+      ? platforms.filter((platform) => platform !== 'web')
+      : platforms;
 
   try {
     if (devServer.isReactServerComponentsEnabled) {
       // In RSC mode, we only need these to be in the client dir.
-      // TODO: Merge back with other copy after we add HMR.
-      await copyPublicFolderAsync(publicPath, path.join(outputPath, 'client'));
+      // TODO: Merge back with other copy after we add SSR.
+      try {
+        await copyPublicFolderAsync(publicPath, path.join(outputPath, 'client'));
+      } catch (error) {
+        Log.error('Failed to copy public directory to dist directory');
+        throw error;
+      }
     } else {
       // NOTE(kitten): The public folder is currently always copied, regardless of targetDomain
       // split. Hence, there's another separate `copyPublicFolderAsync` call below for `web`
@@ -261,14 +268,13 @@ export async function exportAppAsync(
 
       if (devServer.isReactServerComponentsEnabled) {
         const isWeb = platforms.includes('web');
-        if (!(isWeb && useServerRendering)) {
-          await exportApiRoutesStandaloneAsync(devServer, {
-            files,
-            platform: 'web',
-            apiRoutesOnly: !isWeb,
-            templateHtml,
-          });
-        }
+
+        await exportApiRoutesStandaloneAsync(devServer, {
+          files,
+          platform: 'web',
+          apiRoutesOnly: !isWeb,
+          templateHtml,
+        });
       }
 
       // TODO: Use same asset system across platforms again.
@@ -332,7 +338,10 @@ export async function exportAppAsync(
             targetDomain: 'client',
           });
         }
-      } else {
+      } else if (
+        // TODO: Support static export with RSC.
+        !devServer.isReactServerComponentsEnabled
+      ) {
         await exportFromServerAsync(projectRoot, devServer, {
           mode,
           files,